fix: view extension View+ 파일로 수정
